refactor: split lemma_discovery.rs (3079 lines) into a module directory
Pure mechanical relocation — no logic changes. The single 3079-line file
became `lemma_discovery/` with cohesive submodules:

  mod.rs        — shared consts, core types (Binder/TermNode/Conjecture/
                  LawDiscovery/…), pub-API re-exports, #[cfg(test)] tests
  enumerate.rs  — run_discovery + the equational term/conjecture enumerator
  vm_filter.rs  — VM-based counterexample filtering + value sampling
  render.rs     — Lean conjecture rendering, homomorphism check, report/rank
  bricks.rs     — structural lemma bricks (counted-repeat / encoder /
                  relational / monoidal / field-invariant detectors+renderers)

Only visibility keywords and module plumbing (`use super::*;`, `mod`,
`pub use`) changed; every function body is verbatim. The public surface
(`run_discovery`, `vm_filter`, `structural_lemma_groups`, `render_report`,
`rank_candidate_indices`, `lean_lemma_theorem`, `discovery_surface_hash`,
`LawDiscovery`) is re-exported from mod.rs so `codegen::lemma_discovery::*`
paths in commands.rs resolve unchanged.

Verified pure move: the fn/struct/enum/const inventory is byte-identical
(132 items), a normalized content diff shows zero removed/changed lines
(only re-export plumbing added), build + clippy + fmt clean, 16 inline
discovery tests and the full proof_spec suite (95) green.
